requirements.txt
timm
thop
torchtoolbox
